refactor: move convertToWAV to shared audio_utils.go
